Clean up code structure
